+++ Rakefile
@@ -1,22 +1,60 @@
+require 'rubygems'
+require 'bundler'
+begin
+  Bundler.setup(:default, :development)
+rescue Bundler::BundlerError => e
+  $stderr.puts e.message
+  $stderr.puts "Run `bundle install` to install missing gems"
+  exit e.status_code
+end
 require 'rake'
+
+require 'jeweler'
+require './lib/render_component/version.rb'
+
+Jeweler::Tasks.new do |gem|
+  # gem is a Gem::Specification... see http://docs.rubygems.org/read/chapter/20 for more options
+  gem.name = "render_component"
+  gem.version = RenderComponent::Version::STRING
+  gem.homepage = "http://github.com/vhochstein/render_component"
+  gem.license = "MIT"
+  gem.summary = %Q{render actions in other controllers for their rendered response}
+  gem.description = %Q{Components allow you to call other actions for their rendered response while executing another action}
+  gem.email = "david@loudthinking.com"
+  gem.authors = ["David Heinemeier Hansson"]
+  gem.add_runtime_dependency 'railties', '~>3.0.0'
+  # Include your dependencies below. Runtime dependencies are required when using your gem,
+  # and development dependencies are only needed for development (ie running rake tasks, tests, etc)
+  #  gem.add_runtime_dependency 'jabber4r', '> 0.1'
+  #  gem.add_development_dependency 'rspec', '> 1.2.3'
+end
+Jeweler::RubygemsDotOrgTasks.new
+
+
 require 'rake/testtask'
-require 'rake/rdoctask'
- 
-desc 'Default: run unit tests.'
-task :default => :test
- 
 desc 'Test the components plugin.'
 Rake::TestTask.new(:test) do |t|
   t.libs << 'lib'
   t.pattern = 'test/**/*_test.rb'
   t.verbose = true
 end
- 
+
+require 'rcov/rcovtask'
+Rcov::RcovTask.new do |test|
+  test.libs << 'test'
+  test.pattern = 'test/**/test_*.rb'
+  test.verbose = true
+end
+
+task :default => :test
+
+require 'rake/rdoctask'
+
 desc 'Generate documentation for the components plugin.'
 Rake::RDocTask.new(:rdoc) do |rdoc|
+  version = File.exist?('VERSION') ? File.read('VERSION') : ""
   rdoc.rdoc_dir = 'rdoc'
-  rdoc.title = 'Components'
-  rdoc.options << '--line-numbers' << '--inline-source'
-  rdoc.rdoc_files.include('README')
+  rdoc.title = "render_component #{version}"
+  rdoc.rdoc_files.include('README*')
   rdoc.rdoc_files.include('lib/**/*.rb')
 end
