+++ lib/render_component.rb
@@ -1,0 +1,3 @@
+require 'render_component/components'
+require 'action_dispatch/middleware/flash'
+ActionController::Base.send :include, RenderComponent::Components
