+++ lib/render_component/version.rb
@@ -1,0 +1,9 @@
+module RenderComponent
+  module Version
+    MAJOR = 3
+    MINOR = 0
+    PATCH = 1
+
+    STRING = [MAJOR, MINOR, PATCH].compact.join('.')
+  end
+end
